Fixed a couple of linux file issues. Removed operator& overloads from a couple of places where they shouldn't have been done. Fixed a few minor win32 compile issues. Improved some internal COM object wrappers.

+++ src/audio/audio_file_formats/juce_QuickTimeAudioFormat.cpp
@@ -256,7 +256,7 @@ public:
                 }
             }
 
-            int framesToDo = jmin (numSamples, bufferList->mBuffers[0].mDataByteSize / inputStreamDesc.mBytesPerFrame);
+            int framesToDo = jmin (numSamples, (int) (bufferList->mBuffers[0].mDataByteSize / inputStreamDesc.mBytesPerFrame));
             bufferList->mBuffers[0].mDataByteSize = inputStreamDesc.mBytesPerFrame * framesToDo;
 
             UInt32 outFlags = 0;
